[RFC - please comment]  track stable branches instead of hashes.
We want the stable branch, but do we really want to maintain the exact commit?
Tracking the stable branches is easier and we inherit the fixes w/o changing our manifest again.

Please comment. If this is fine please merge.

Especially this should be changed:
<project name="meta-openembedded" path="meta-openembedded" remote="openembedded" revision="9efaed99125b1c4324663d9a1b2d3319c74e7278" />

otherwise we are  "147 files changed, 7967 insertions(+), 888 deletions(-)"
behind dizzy branch for meta-openembedded !
Multiple CVEs were fixed in the branch already.

Change-Id: I31689d8ea49d8c876be8bb5608df89ad4c459ab9
Signed-off-by: Jan-Simon Möller <jsmoeller@linuxfoundation.org>