Update patch set 1
Patch Set 1: Code-Review-1

I basically agree to you.

Before to remove hashes, we should define integration and release policy like stable/snapshot/nightly and so on and prepare branches/manifests for them if needed.
The reason is because it sometimes would fail to build with heads of upstream, or because it sometimes would be difficult to analyze the issues depending on each environments like when you did 'repo sync'.

Patch-set: 1