Update patch set 1
Patch Set 1:

> @jsm, Shouldnt bots find this problem before uses?

Well, dependency problems can be heisenbugs. glib-2.0-native might be pulled-
in by another package and be available 'just-in-time'.

> I also see different dates in:
> https://download.automotivelinux.org/AGL/snapshots/master/latest/
Those are the nightly builds, not the CI builds. The daily snapshots are b/c 
not all boards build successfully and on-top we had jenkins issues during the 
last 2 days.

> Is this intentional?
Yes. We always link to the latest built snapshot.

Patch-set: 1