Update patch set 2
Patch Set 2:

I'm not sure if this really creates a reproducable version.
Let's take HomeScreen app as an example. chinnok_2.99.2.xml defines a fix revision for the mate-agl-demo layer. But inside, there is a homescreen_git.bb recipe. This will always sync the latest git revision.
And I suspect, there are other recipes that will do so...

Patch-set: 2