Update patch set 1
Patch Set 1: Code-Review-1

Is adding meta-linaro required for optee to boot gen3 board?
If so, please describe this in the commit message.

I think this patch has 2 features, 1st is update revision of meta-renesas to morty and 2nd is adding meta-linaro, so the subject doesn't represent whole purpose.

Patch-set: 1
Label: Code-Review=-1