Update patch set 3
Patch Set 3:

Ok, I need to play the bad guy once more:

- https://layers.openembedded.org/layerindex/branch/master/recipes/?q=python-pycrypto
-- we already have python-pycrypto in meta-agl, why a 2nd one - what is the diff ?

- so this leaves only one single recipe 'python-wand' on the list why we pull-in a whole new repo ?!

- We're currently not using optee as per your link to the device template due to the BBMASK
-- https://gerrit.automotivelinux.org/gerrit/gitweb?p=AGL/meta-renesas-rcar-gen3.git;a=blob;f=meta-rcar-gen3/docs/sample/conf/h3ulcb/poky-gcc/bsp/bblayers.conf;h=3b2c068aefb8f5f5093676978cea81a750f308cd;hb=8f35bd31157eb3e62c181102afedb8cb730e21ea
-- http://www.yoctoproject.org/docs/latest/mega-manual/mega-manual.html#var-BBMASK

So effectively, we have 'just' one recipe left and that is 'python-wand'. This could be added to upstream meta-python (which we include already), which is the go-to for python recipes anyway, and we're done.

My main concern is that we're pulling a lot of recipes into the codebase that we don't need or use.

Patch-set: 3