Update patch set 1
Patch Set 1:

Hi Martin,

> Have you thought about contributing the recipes here to upstream meta-openembedded?

Thank you for your suggestion. I think meta-openembedded is supported for recipes which are generic in embedded Linux system. SmartDeviceLink is a specific software component for IVI system. A separate layer is reality.

>  I think it would be helpful to get broader usage and also have easier maintenance in making sure the recipes work well in the broader universe of OE.

Actually the meta layer is on GitHub. So it's easy to develop among community.

Regards,
Phong.

Patch-set: 1