Update patch set 1
Patch Set 1:

There was a mismatch in the git repository... The tag 4.99.1 was set on the tip of master while the manifest pointed to the branch pyro_bsp_2.23 which was 2 commits behind the tagged rev.
So I realigned everything: pyro_bsp_2.23 is now synced on master and thus all refs (master or pyro_bsp_2.23 or eel_4.99.1) point to the same revision.
I think that keeping the branch "pyro_bsp_2.23" is more explicit than "master": when bumping the bsp or the poky version, it's easier to create a new branch and name it accordingly. But for CI, it's an issue as builds happen only if something is pushed to master (could this be changed ?).
In fact for this repo which is a combination of upstream repos + AGL specific patches to align on poky version, we shouldn't have any master branch...
Any ideas to make things better here ?

Patch-set: 1