Update patch set 1
Patch Set 1: Code-Review+1

It is not part of meta-agl-demo at all. meta-agl-demo is the IVI UI
and its components. The profiles and layers are setup so the cluster
is its own toplevel project.
We can debate about devel, but to me it is a new toplevel/image
project  (and not a platform component) we can justify the
meta-agl-cluster-demo .

Patch-set: 1
Label: Code-Review=+1