Update to latest zeus commits
Update upstream repositories to latest zeus commits, re-enabling a few
that had previously been disabled for lack of zeus branches.

Note that the upstream Yocto Project repos have specifically not been
set to the 3.0.1 commits since 3.0.2 is releasing soon and it seems
better to test with a level close to it in preparation.

Change-Id: I4eb5f01537c95cdd9774e72b7a27082c41a4e9e1
Signed-off-by: Scott Murray <scott.murray@konsulko.com>