Update patch set 2
Patch Set 2:

The reason to pull this in asap is that there will be brownouts where the github servers will not be available through the 'git protocol' even before the january deadline. To avoid this, we backport the fix now and include the next stable release when available.

Patch-set: 2
Reviewer: Gerrit User 4 <4@2db66874-6d0c-4434-a3ad-56d904acc56e>