Update patch set 1
Patch Set 1:

I did see this in the 2 CI test runs for the bbe. E.g.
https://lava.automotivelinux.org/scheduler/job/24019#L635 

It seems just a warning and it does not seems to affect the results in the end. 
NBD is used in CI (aka we tftp-boot and then instead of NFS we do use NBD).

Patch-set: 1
Reviewer: Gerrit User 4 <4@2db66874-6d0c-4434-a3ad-56d904acc56e>