Update patch set 1
Patch Set 1: Code-Review+1

> Patch Set 1:
> 
> dunfell-dev follows latest YP3.1.x branch, but dunfell branch of meta-renesas is fixed(YP3.1.11)

Okay, though that should perhaps be explained in the README.md in meta-renesas or meta-renesas/meta-rcar-gen3, as it is counter-intuitive in my opinion.  If you're able to talk to the BSP maintainers I notice that the dunfell-dev branch (and the dunfell branch as well) currently does not build against latest meta-openembedded dunfell; the linuxptp_2.0.bbappend in meta-rcar-gen3 needs to be renamed to handle a recent bump to 2.0.1 in upstream.  It would be good if that could be fixed so we do not have to pin meta-openembedded at an old commit when we bump things for the upgrade to 3.1.15.

Patch-set: 1
Reviewer: Gerrit User 597 <597@2db66874-6d0c-4434-a3ad-56d904acc56e>
Label: Code-Review=+1