Update Rust mixin layer
Switch from kirkstone/rust-1.68 branch in meta-lts-mixin to new
kirkstone/rust-1.70 branch to future proof building the KUKSA.val
databroker and the various vhost device backends.  The target
directory in external has been changed to drop the version number
so that the configuration template will not need to be changed
again if for some reason we do need to do another bump.

Bug-AGL: SPEC-5060

Change-Id: Id3d6c3556bb1fc3dc0aad2db4ab35b44b5bbe571
Signed-off-by: Scott Murray <scott.murray@konsulko.com>